改完文件，神功大成

+++ code/data_list.py
@@ -108,7 +108,7 @@ def make_dataset(image_list, labels):
         if len(image_list[0].split()) > 2:
             images = [(val.split()[0], np.array([int(la) for la in val.split()[1:]])) for val in image_list]
         else:
-            images = [(val.split()[0], int(val.split()[1])) for val in image_list]
+            images = [(val.split()[0], float(val.split()[1])) for val in image_list]
     return images
